The camera has been modified: Variable EyePosition: The camera position vector - The camera position's x- and y-coordinates match those of the second instance of the object, even as it moves. - The camera position's z-coordinate differs from those of the second instance of the object, and can be manually adjusted by keyboard keys.
Variable FocusPosition: The focal point position vector
- The camera points at the second instance of the object even as it moves. Thus the second instance of the object appears static, while the first instance of the object (which is static) appears to move in the direction opposite to how the first instance of the object moves.

Variable UpDirection: The up direction vector
- The up direction vector is a unit vector that points in the positive y direction, which is the top of the camera.

If EyePosition and FocusPosition become equal this is an error condition. Therefore if this occurs EyePosition is slightly adjusted to keep them unequal as required (previously this program terminated with a return value indicating the error condition).

The call to the RenderText function now results in displaying dynamic text: It displays the values of key variables.

Comments have been improved.