The OBJECT 'named structure' specifies the set of attributes of a 3D object.  It has been modified to contain vertex, index, constant buffer, and other attributes.
Each element of OurObjects, the array of OBJECT structures, specifies the set of attributes of one of multiple 3D objects, including attributes specified by a corresponding Wavefront .obj file.

Each element of OurObjects is processed for rendering a 3D object.

Comments have been changed and added in support of the above.