Create a device that supports the debug layer when the program starts with debugging enabled (Visual Studio menu/ Debug/ Start Debugging).
Changed the calculations of Angle and Angle2 to wrap within the range 0 - 359 instead of continuously increasing (continuously increasing a large number by a small amount results in eventually attaining a large number that, due to loss of precision, does not change in response to being incremented).

Reordered some statements and some function parameters.

Modified some comments.